Change the discrepancies between the Java Persistence layer and database

+++ src/main/java/io/kovin/dispatch/management/system/model/request/load/CreateLoadLocationRequest.java
@@ -8,7 +8,7 @@ public record CreateLoadLocationRequest(
     LocalDate date,
     LocalTime time,
     String location,
-    Integer order,
+    Short order,
     String address,
     String timezone
 ) {
